requirements.txt
python
pytest
pytest-playwright
